Refactored the files for the GBA formats
