+++ frontend/src/components/ConvertGbaActionReplay.vue
@@ -6,23 +6,21 @@
           <b-row no-gutters align-h="center" align-v="start">
             <b-col cols=12>
               <b-jumbotron fluid :header-level="$mq | mq({ xs: 5, sm: 5, md: 5, lg: 5, xl: 4 })">
-                <template v-slot:header>GameShark{{'\xa0'}}SP</template> <!-- Not sure why &nbsp; doesn't work here, but this does -->
+                <template v-slot:header>Action Replay</template>
               </b-jumbotron>
             </b-col>
           </b-row>
-          <div v-if="this.conversionDirection === 'convertToEmulator'" class="inputs-row">
+          <div v-if="this.conversionDirection === 'convertToEmulator'">
             <input-file
-              id="choose-gameshark-file"
-              @load="readGameSharkSaveData($event)"
+              @load="readActionReplaySaveData($event)"
               :errorMessage="this.errorMessage"
-              placeholderText="Choose a file to convert (*.gsv)"
-              acceptExtension=".gsv"
+              placeholderText="Choose a file to convert (*.xps)"
+              acceptExtension=".xps"
               :leaveRoomForHelpIcon="false"
             />
           </div>
-          <div v-else class="inputs-row">
-            <output-filename v-model="outputFilename" :leaveRoomForHelpIcon="true"/>
-            <output-filesize v-model="outputFilesize" id="gameshark-filesize"/>
+          <div v-else>
+            <output-filename v-model="outputFilename" :leaveRoomForHelpIcon="false"/>
           </div>
         </b-col>
         <b-col sm=12 md=2 lg=2 xl=2 align-self="start">
@@ -30,13 +28,8 @@
             :horizontalLayout="['md', 'lg', 'xl']"
             :verticalLayout="['xs', 'sm']"
             :conversionDirection="this.conversionDirection"
-            disableDirection="convertToRetron5"
             @change="changeConversionDirection($event)"
-            id="conversion-direction"
           />
-          <b-popover target="conversion-direction" triggers="focus hover" placement="bottom">
-            Not enough is known about GameShark SP saves to be able to write out a file in this format. We can currently only extract the raw save from these files.
-          </b-popover>
         </b-col>
         <b-col sm=12 md=5 align-self="start">
           <b-row no-gutters align-h="center" align-v="start">
@@ -46,27 +39,15 @@
               </b-jumbotron>
             </b-col>
           </b-row>
-          <div v-if="this.conversionDirection === 'convertToEmulator'" class="inputs-row">
-            <output-filename v-model="outputFilename" :leaveRoomForHelpIcon="true"/>
-            <output-filesize v-model="outputFilesize" id="raw-filesize"/>
+          <div v-if="this.conversionDirection === 'convertToEmulator'">
+            <output-filename v-model="outputFilename" :leaveRoomForHelpIcon="false"/>
           </div>
-          <div v-else class="inputs-row">
+          <div v-else>
             <input-file
-              id="choose-raw-file"
               @load="readEmulatorSaveData($event)"
               :errorMessage="this.errorMessage"
               placeholderText="Choose a file to convert"
-              :leaveRoomForHelpIcon="true"
-            />
-            <input-file
-              id="choose-raw-file-rom"
-              @load="readRomData($event)"
-              :errorMessage="null"
-              placeholderText="Choose the ROM for this file (*.gba)"
-              helpText="GameShark SP save files contain some information from the corresponding ROM, and some emulators check this information before allowing the save to be loaded.
-              All processing by this website is done on your local machine, and your ROMs are not sent anywhere."
-              acceptExtension=".gba"
-              :leaveRoomForHelpIcon="true"
+              :leaveRoomForHelpIcon="false"
             />
           </div>
         </b-col>
@@ -74,10 +55,10 @@
       <b-row class="justify-content-md-center" align-h="center">
         <b-col cols="auto" sm=4 md=3 lg=2 align-self="center">
           <b-button
-            class="gameshark-sp-convert-button"
+            class="action-replay-convert-button"
             variant="success"
             block
-            :disabled="!this.gameSharkSaveData || !outputFilename"
+            :disabled="!this.actionReplaySaveData || !outputFilename"
             @click="convertFile()"
           >
           Convert!
@@ -87,7 +68,7 @@
       <b-row>
         <b-col>
           <div class="help">
-            Help: how do I&nbsp;<b-link href="https://www.manualslib.com/manual/515605/Mad-Catz-Gameshark-Sp.html?page=17#manual">copy save files to and from my GameShark SP</b-link>?
+            Help: how do I copy files to and from my <b-link href="https://www.youtube.com/watch?v=o7WQYnYYT6Y">Action Replay DSi</b-link> or <b-link href="http://uk.codejunkies.com/downloads/manuals/armax_gba_ds_english.pdf">Action Replay MAX DUO</b-link>?
           </div>
           <div class="help">
             Help: what devices can copy save files to and from a GBA cartridge?<br><router-link to="/retron-5">Retron{{'\xa0'}}5</router-link> or <b-link href="https://www.epilogue.co/product/gb-operator">GB{{'\xa0'}}Operator</b-link> or <b-link href="https://bennvenn.myshopify.com/products/usb-gb-c-cart-dumper-the-joey-jr">Joey{{'\xa0'}}Jr.</b-link> or <b-link href="https://www.gbxcart.com/">GBxCart{{'\xa0'}}RW</b-link>
@@ -106,79 +87,67 @@
 
 <style scoped>
 
-/* Separate class for each different button to enable tracking in google tag manager */
-.gameshark-sp-convert-button {
+.action-replay-convert-button {
   margin-top: 1em;
 }
 
 .help {
   margin-top: 1em;
 }
-
-.inputs-row {
-  min-height: 4.8em;
-}
-
 </style>
 
 <script>
-import { BPopover } from 'bootstrap-vue';
 import { saveAs } from 'file-saver';
 import Util from '../util/util';
 import InputFile from './InputFile.vue';
 import OutputFilename from './OutputFilename.vue';
-import OutputFilesize from './OutputFilesize.vue';
 import ConversionDirection from './ConversionDirection.vue';
-import GameSharkSpSaveData from '../save-formats/GameSharkSP';
+import ActionReplaySaveData from '../save-formats/GBA/ActionReplay';
 
 export default {
-  name: 'ConvertGameSharkSP',
+  name: 'ConvertGbaActionReplay',
   data() {
     return {
-      gameSharkSaveData: null,
-      romData: null,
-      emulatorSaveData: null,
-      emulatorSaveDataFilename: null,
+      actionReplaySaveData: null,
       errorMessage: null,
       outputFilename: null,
-      outputFilesize: null,
       conversionDirection: 'convertToEmulator',
     };
   },
   components: {
-    BPopover,
     ConversionDirection,
     InputFile,
     OutputFilename,
-    OutputFilesize,
   },
   methods: {
     changeConversionDirection(newDirection) {
       this.conversionDirection = newDirection;
-      this.gameSharkSaveData = null;
+      this.actionReplaySaveData = null;
       this.errorMessage = null;
       this.outputFilename = null;
-      this.outputFilesize = null;
     },
-    readGameSharkSaveData(event) {
+    readActionReplaySaveData(event) {
       this.errorMessage = null;
       try {
-        this.gameSharkSaveData = GameSharkSpSaveData.createFromGameSharkSpData(event.arrayBuffer);
+        this.actionReplaySaveData = ActionReplaySaveData.createFromActionReplayData(event.arrayBuffer);
         this.outputFilename = Util.changeFilenameExtension(event.filename, 'srm');
-        this.outputFilesize = this.gameSharkSaveData.getRawSaveData().byteLength;
       } catch (e) {
         this.errorMessage = e.message;
-        this.gameSharkSaveData = null;
-        this.outputFilename = null;
-        this.outputFilesize = null;
+        this.actionReplaySaveData = null;
+      }
+    },
+    readEmulatorSaveData(event) {
+      this.errorMessage = null;
+      try {
+        this.actionReplaySaveData = ActionReplaySaveData.createFromEmulatorData(event.arrayBuffer);
+        this.outputFilename = Util.changeFilenameExtension(event.filename, 'xps');
+      } catch (e) {
+        this.errorMessage = e.message;
+        this.actionReplaySaveData = null;
       }
     },
     convertFile() {
-      let outputArrayBuffer = this.gameSharkSaveData.getRawSaveData();
-
-      if (outputArrayBuffer.byteLength !== this.outputFilesize) {
-        outputArrayBuffer = Util.resizeRawSave(outputArrayBuffer, this.outputFilesize);
-      }
+      const outputArrayBuffer = (this.conversionDirection === 'convertToEmulator') ? this.actionReplaySaveData.getRawSaveData() : this.actionReplaySaveData.getArrayBuffer();
 
       const outputBlob = new Blob([outputArrayBuffer], { type: 'application/octet-stream' });
