+++ frontend/src/components/ConvertGbaGameShark.vue
@@ -6,21 +6,23 @@
           <b-row no-gutters align-h="center" align-v="start">
             <b-col cols=12>
               <b-jumbotron fluid :header-level="$mq | mq({ xs: 5, sm: 5, md: 5, lg: 5, xl: 4 })">
-                <template v-slot:header>Action Replay</template>
+                <template v-slot:header>GameShark</template>
               </b-jumbotron>
             </b-col>
           </b-row>
-          <div v-if="this.conversionDirection === 'convertToEmulator'">
+          <div v-if="this.conversionDirection === 'convertToEmulator'" class="inputs-row">
             <input-file
-              @load="readActionReplaySaveData($event)"
+              id="choose-gameshark-file"
+              @load="readGameSharkSaveData($event)"
               :errorMessage="this.errorMessage"
-              placeholderText="Choose a file to convert (*.xps)"
-              acceptExtension=".xps"
+              placeholderText="Choose a file to convert (*.sps)"
+              acceptExtension=".sps"
               :leaveRoomForHelpIcon="false"
             />
           </div>
-          <div v-else>
-            <output-filename v-model="outputFilename" :leaveRoomForHelpIcon="false"/>
+          <div v-else class="inputs-row">
+            <output-filename v-model="outputFilename" :leaveRoomForHelpIcon="true"/>
+            <output-filesize v-model="outputFilesize" id="gameshark-filesize"/>
           </div>
         </b-col>
         <b-col sm=12 md=2 lg=2 xl=2 align-self="start">
@@ -39,15 +41,27 @@
               </b-jumbotron>
             </b-col>
           </b-row>
-          <div v-if="this.conversionDirection === 'convertToEmulator'">
-            <output-filename v-model="outputFilename" :leaveRoomForHelpIcon="false"/>
+          <div v-if="this.conversionDirection === 'convertToEmulator'" class="inputs-row">
+            <output-filename v-model="outputFilename" :leaveRoomForHelpIcon="true"/>
+            <output-filesize v-model="outputFilesize" id="raw-filesize"/>
           </div>
-          <div v-else>
+          <div v-else class="inputs-row">
             <input-file
+              id="choose-raw-file"
               @load="readEmulatorSaveData($event)"
               :errorMessage="this.errorMessage"
               placeholderText="Choose a file to convert"
-              :leaveRoomForHelpIcon="false"
+              :leaveRoomForHelpIcon="true"
+            />
+            <input-file
+              id="choose-raw-file-rom"
+              @load="readRomData($event)"
+              :errorMessage="null"
+              placeholderText="Choose the ROM for this file (*.gba)"
+              helpText="GameShark save files contain some information from the corresponding ROM, and some emulators check this information before allowing the save to be loaded.
+              All processing by this website is done on your local machine, and your ROMs are not sent anywhere."
+              acceptExtension=".gba"
+              :leaveRoomForHelpIcon="true"
             />
           </div>
         </b-col>
@@ -55,10 +69,10 @@
       <b-row class="justify-content-md-center" align-h="center">
         <b-col cols="auto" sm=4 md=3 lg=2 align-self="center">
           <b-button
-            class="action-replay-convert-button"
+            class="gameshark-convert-button"
             variant="success"
             block
-            :disabled="!this.actionReplaySaveData || !outputFilename"
+            :disabled="!this.gameSharkSaveData || !outputFilename"
             @click="convertFile()"
           >
           Convert!
@@ -68,7 +82,7 @@
       <b-row>
         <b-col>
           <div class="help">
-            Help: how do I copy files to and from my <b-link href="https://www.youtube.com/watch?v=o7WQYnYYT6Y">Action Replay DSi</b-link> or <b-link href="http://uk.codejunkies.com/downloads/manuals/armax_gba_ds_english.pdf">Action Replay MAX DUO</b-link>?
+            Help: how do I&nbsp;<b-link href="https://gamehacking.org/vb/forum/video-game-hacking-and-development/school-of-hacking/14153-gameshark-advance-and-color-save-backup">copy save files to and from my GameShark</b-link>?
           </div>
           <div class="help">
             Help: what devices can copy save files to and from a GBA cartridge?<br><router-link to="/retron-5">Retron{{'\xa0'}}5</router-link> or <b-link href="https://www.epilogue.co/product/gb-operator">GB{{'\xa0'}}Operator</b-link> or <b-link href="https://bennvenn.myshopify.com/products/usb-gb-c-cart-dumper-the-joey-jr">Joey{{'\xa0'}}Jr.</b-link> or <b-link href="https://www.gbxcart.com/">GBxCart{{'\xa0'}}RW</b-link>
@@ -87,30 +101,42 @@
 
 <style scoped>
 
-.action-replay-convert-button {
+/* Separate class for each different button to enable tracking in google tag manager */
+.gameshark-convert-button {
   margin-top: 1em;
 }
 
 .help {
   margin-top: 1em;
 }
+
+.inputs-row {
+  min-height: 4.8em;
+}
+
 </style>
 
 <script>
+import dayjs from 'dayjs';
 import { saveAs } from 'file-saver';
 import Util from '../util/util';
 import InputFile from './InputFile.vue';
 import OutputFilename from './OutputFilename.vue';
+import OutputFilesize from './OutputFilesize.vue';
 import ConversionDirection from './ConversionDirection.vue';
-import ActionReplaySaveData from '../save-formats/ActionReplay';
+import GameSharkSaveData from '../save-formats/GBA/GameShark';
 
 export default {
-  name: 'ConvertActionReplay',
+  name: 'ConvertGbaGameShark',
   data() {
     return {
-      actionReplaySaveData: null,
+      gameSharkSaveData: null,
+      romData: null,
+      emulatorSaveData: null,
+      emulatorSaveDataFilename: null,
       errorMessage: null,
       outputFilename: null,
+      outputFilesize: null,
       conversionDirection: 'convertToEmulator',
     };
   },
@@ -118,36 +144,66 @@ export default {
     ConversionDirection,
     InputFile,
     OutputFilename,
+    OutputFilesize,
   },
   methods: {
     changeConversionDirection(newDirection) {
       this.conversionDirection = newDirection;
-      this.actionReplaySaveData = null;
+      this.gameSharkSaveData = null;
+      this.romData = null;
+      this.emulatorSaveData = null;
+      this.emulatorSaveDataFilename = null;
       this.errorMessage = null;
       this.outputFilename = null;
+      this.outputFilesize = null;
     },
-    readActionReplaySaveData(event) {
-      this.errorMessage = null;
-      try {
-        this.actionReplaySaveData = ActionReplaySaveData.createFromActionReplayData(event.arrayBuffer);
-        this.outputFilename = Util.changeFilenameExtension(event.filename, 'srm');
-      } catch (e) {
-        this.errorMessage = e.message;
-        this.actionReplaySaveData = null;
-      }
+    readRomData(event) {
+      this.romData = event.arrayBuffer;
+      this.tryToCreateGameSharkSaveDataFromEmulatorSaveData();
     },
     readEmulatorSaveData(event) {
+      this.emulatorSaveData = event.arrayBuffer;
+      this.emulatorSaveDataFilename = event.filename;
+      this.tryToCreateGameSharkSaveDataFromEmulatorSaveData();
+    },
+    readGameSharkSaveData(event) {
       this.errorMessage = null;
       try {
-        this.actionReplaySaveData = ActionReplaySaveData.createFromEmulatorData(event.arrayBuffer);
-        this.outputFilename = Util.changeFilenameExtension(event.filename, 'xps');
+        this.gameSharkSaveData = GameSharkSaveData.createFromGameSharkData(event.arrayBuffer);
+        this.outputFilename = Util.changeFilenameExtension(event.filename, 'srm');
+        this.outputFilesize = this.gameSharkSaveData.getRawSaveData().byteLength;
       } catch (e) {
         this.errorMessage = e.message;
-        this.actionReplaySaveData = null;
+        this.gameSharkSaveData = null;
+        this.outputFilename = null;
+        this.outputFilesize = null;
+      }
+    },
+    tryToCreateGameSharkSaveDataFromEmulatorSaveData() {
+      if ((this.romData !== null) && (this.emulatorSaveData !== null) && (this.emulatorSaveDataFilename !== null)) { // Need to be careful that the user can choose the save data and ROM files in any order, so only proceed once both have been chosen
+        this.errorMessage = null;
+
+        try {
+          const title = Util.removeFilenameExtension(this.emulatorSaveDataFilename);
+          const date = dayjs().format('DD/MM/YYYY hh:mm:ss a');
+          const notes = 'Created with savefileconverter.com';
+          this.gameSharkSaveData = GameSharkSaveData.createFromEmulatorData(this.emulatorSaveData, title, date, notes, this.romData);
+          this.outputFilename = this.changeFilenameExtension(this.emulatorSaveDataFilename, 'sps');
+          this.outputFilesize = this.gameSharkSaveData.getRawSaveData().byteLength;
+        } catch (e) {
+          this.errorMessage = e.message;
+          this.gameSharkSaveData = null;
+          this.outputFilename = null;
+          this.outputFilesize = null;
+        }
       }
     },
     convertFile() {
-      const outputArrayBuffer = (this.conversionDirection === 'convertToEmulator') ? this.actionReplaySaveData.getRawSaveData() : this.actionReplaySaveData.getArrayBuffer();
+      if (this.gameSharkSaveData.getRawSaveData().byteLength !== this.outputFilesize) {
+        this.gameSharkSaveData = GameSharkSaveData.createWithNewSize(this.gameSharkSaveData, this.outputFilesize);
+      }
+
+      const outputArrayBuffer = (this.conversionDirection === 'convertToEmulator') ? this.gameSharkSaveData.getRawSaveData() : this.gameSharkSaveData.getArrayBuffer();
 
       const outputBlob = new Blob([outputArrayBuffer], { type: 'application/octet-stream' });
