--- frontend/src/views/ActionReplay.vue
@@ -1,17 +1,0 @@
-<template>
-  <div class="home">
-    <ConvertActionReplay/>
-  </div>
-</template>
-
-<script>
-// @ is an alias to /src
-import ConvertActionReplay from '@/components/ConvertActionReplay.vue';
-
-export default {
-  name: 'ActionReplay',
-  components: {
-    ConvertActionReplay,
-  },
-};
-</script>
