--- frontend/src/views/GameShark.vue
@@ -1,17 +1,0 @@
-<template>
-  <div class="home">
-    <ConvertGameShark/>
-  </div>
-</template>
-
-<script>
-// @ is an alias to /src
-import ConvertGameShark from '@/components/ConvertGameShark.vue';
-
-export default {
-  name: 'GameShark',
-  components: {
-    ConvertGameShark,
-  },
-};
-</script>
