--- frontend/src/views/GameSharkSP.vue
@@ -1,17 +1,0 @@
-<template>
-  <div class="home">
-    <ConvertGameSharkSP/>
-  </div>
-</template>
-
-<script>
-// @ is an alias to /src
-import ConvertGameSharkSP from '@/components/ConvertGameSharkSP.vue';
-
-export default {
-  name: 'GameSharkSP',
-  components: {
-    ConvertGameSharkSP,
-  },
-};
-</script>
