+++ frontend/src/views/GbaActionReplay.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="home">
+    <ConvertGbaActionReplay/>
+  </div>
+</template>
+
+<script>
+// @ is an alias to /src
+import ConvertGbaActionReplay from '@/components/ConvertGbaActionReplay.vue';
+
+export default {
+  name: 'GbaActionReplay',
+  components: {
+    ConvertGbaActionReplay,
+  },
+};
+</script>
