+++ frontend/src/views/GbaGameSharkSP.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="home">
+    <ConvertGbaGameSharkSP/>
+  </div>
+</template>
+
+<script>
+// @ is an alias to /src
+import ConvertGbaGameSharkSP from '@/components/ConvertGbaGameSharkSP.vue';
+
+export default {
+  name: 'GbaGameSharkSP',
+  components: {
+    ConvertGbaGameSharkSP,
+  },
+};
+</script>
