+++ frontend/tests/unit/save-formats/GBA/GameSharkSP.spec.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
-import GameSharkSpSaveData from '@/save-formats/GameSharkSP';
+import GameSharkSpSaveData from '@/save-formats/GBA/GameSharkSP';
 import ArrayBufferUtil from '#/util/ArrayBuffer';
 
-const DIR = './tests/unit/save-formats/data/gameshark-sp';
+const DIR = './tests/unit/save-formats/data/gba/gameshark-sp';
 
 const GAMESHARKSP_FILENAME = `${DIR}/final-fantasy-tactics-advance.22864.gsv`;
 const RAW_FILENAME = `${DIR}/final-fantasy-tactics-advance.22864.srm`;
